% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clearFilters}
\alias{clearFilters}
\title{Clear all filters from a table}
\usage{
clearFilters(session, tbl, doFilter = TRUE)
}
\arguments{
\item{tbl}{Name of the table to clear.}

\item{doFilter}{Unfilter the table after clearing the filter strings.}

\item{Session}{Shiny session object.}
}
\description{
Clear all filters from a table
}
\examples{
clearFilters(session, "mtcars")
}

