% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableFilter.R
\name{enableEdit}
\alias{enableEdit}
\title{Enable editing of a tableFilter widget}
\usage{
enableEdit(session, tbl, col = NULL)
}
\arguments{
\item{tbl}{Name of the table to be edited.}

\item{col}{Enable selecting of column \code{col} only.}

\item{Session}{Shiny session object.}
}
\description{
Enable editing of a tableFilter widget
}
\examples{
enableEdit(session, "mtcars", "col_0")
}

