% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{rejectEdit}
\alias{rejectEdit}
\title{Give feedback in case of validaton failure.}
\usage{
rejectEdit(session, tbl, row, col, id, value = NULL, color = "red")
}
\arguments{
\item{session}{Shiny session object.}

\item{tbl}{Name of the table beeing edited.}

\item{row}{Row number as received via edit input.}

\item{col}{Column number as received via edit input.}

\item{id}{Unique identifier of the edit event, received via the edit input}

\item{value}{Reset the input to this value if not null.}

\item{color}{Text colour of a failed edit.}
}
\description{
For each input event in a tableFilter widget a message is sent via a shiny 
input. After input validation \code{rejectEdit} can be used to give visual feedback
to the user in case of a validation failure.
}
\details{
In edit mode the a tableFilter widget creates a shiny input element. The name
of this input is the name of the corresponding output element followed by 
"_edit". For each edit event in the html table this input receives a list 
giving a unique ID of the edit event ("id"), the row ("row"), the column 
("col") and the new value ("val") of the cell. Row and column numbers are in 
R coordinates. If \code{showRowNames} is \code{TRUE}, the column number of 
the rownames is 0.
}
