% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{resetInput}
\alias{resetInput}
\title{Reset a D3tf input element}
\usage{
resetInput(session, input)
}
\arguments{
\item{Session}{Shiny session object.}

\item{tbl}{Name of the input (output name  + "_edit").}
}
\description{
Reset a D3tf input element
}
\examples{
resetInput(session, "mtcars_edit")
}
