% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{setCellValue}
\alias{setCellValue}
\title{Set cell value}
\usage{
setCellValue(session, tbl, row, col, value, feedback = FALSE)
}
\arguments{
\item{tbl}{Name of the table.}

\item{row}{Row number (one-based).}

\item{col}{Column number (one-based). If \code{showRowNames == TRUE}, the rownames column is number zero.}

\item{value}{Cell value to set.}

\item{feedback}{Send edit event back to server.}

\item{Session}{Shiny session object.}
}
\description{
Set cell value
}
\examples{
setCellValue(session, "mtcars", row = 8, col = 3, val = 8)
}
