% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{setFootCellValue}
\alias{setFootCellValue}
\title{Set foot cell value}
\usage{
setFootCellValue(session, tbl, row, col, value, feedback = FALSE)
}
\arguments{
\item{tbl}{Name of the table.}

\item{row}{Footer row number (one-based).}

\item{col}{Footer olumn number (one-based). If \code{showRowNames == TRUE}, the rownames column is number zero.}

\item{value}{Cell value to set.}

\item{Session}{Shiny session object.}
}
\description{
Set foot cell value
}
\examples{
setFootCellValue(session, "mtcars", row = 1, col = 1, val = 8)
}
