% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{setRowClass}
\alias{setRowClass}
\title{Highlight a row using bootstrap classes}
\usage{
setRowClass(session, tbl, row, class)
}
\arguments{
\item{tbl}{Name of the table.}

\item{row}{Number of the row to color.}

\item{class}{Bootstrap contextual class (\code{"active"}, \code{"info"}, \code{"success"}, \code{"warning"}, or \code{"danger"}). \code{"none"} removes the highlighting. \code{"info"} is reserved for selected rows.}

\item{Session}{Shiny session object.}
}
\description{
Highlight a row using bootstrap classes
}
\examples{
setRowClass(session, "mtcars", 3, "success")
}
