% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableFilter.R
\name{clearFilters}
\alias{clearFilters}
\title{Clear all filters from a table}
\usage{
clearFilters(session, tbl, doFilter = TRUE)
}
\arguments{
\item{tbl}{Name of the table to clear.}

\item{Session}{Shiny session object.}
}
\description{
Clear all filters from a table
}
\examples{
clearFilters(session, "mtcars")
}

