% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableFilter.R
\name{disableEdit}
\alias{disableEdit}
\title{Disable editing of a tableFilter widget}
\usage{
disableEdit(session, tbl, col = NULL)
}
\arguments{
\item{tbl}{Name of the table to disable editing.}

\item{col}{Disable selecting of column \code{col} only.}

\item{Session}{Shiny session object.}
}
\description{
Disable editing of a tableFilter widget
}
\examples{
disableEdit(session, "mtcars", "col_0")
}

