% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3TableFilter.R
\name{disableEdit}
\alias{disableEdit}
\title{Disable editing of a tableFilter widget}
\usage{
disableEdit(session, tbl, cols = NULL)
}
\arguments{
\item{tbl}{Name of the table to disable editing.}

\item{cols}{Disable editing of single column (\code{"col_0"}) or multiple columns (\code{c("col_0", "col_1")}).}

\item{Session}{Shiny session object.}
}
\description{
Disable editing of a tableFilter widget
}
\examples{
disableEdit(session, "mtcars", c("col_1", "col_2"))
}
