% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableFilter.R
\name{setCellValue}
\alias{setCellValue}
\title{Set cell value}
\usage{
setCellValue(session, tbl, row, col, value)
}
\arguments{
\item{tbl}{Name of the table.}

\item{row}{Row number (one-based).}

\item{col}{Column number (one-based). If \code{showRowNames == TRUE}, the rownames column is number zero.}

\item{Session}{Shiny session object.}

\item{val}{Cell value to set.}
}
\description{
Set cell value
}
\examples{
setCellValue(session, "mtcars", 3, "success")
}

